Refactor: useNewProduct분리
